modifying profile model to just be profiledata
